+++ build.gradle
@@ -27,6 +27,11 @@ test {
     }
 }
 
+dependencies {
+    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.5.0'
+    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.5.0'
+}
+
 application {
     mainClassName = "seedu.duke.Duke"
 }
