+++ src/main/java/duke/Duke.java
@@ -2,7 +2,10 @@ package duke;
 
 import duke.command.Command;
 import duke.exception.DukeException;
-import duke.exception.DukeFileNotFoundException;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
