+++ src/main/java/duke/command/DeleteCommand.java
@@ -3,9 +3,9 @@ package duke.command;
 import duke.exception.DukeException;
 import duke.exception.DukeOutOfBoundsException;
 import duke.task.Task;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 import java.io.IOException;
