+++ src/main/java/duke/command/TodoCommand.java
@@ -1,10 +1,9 @@
 package duke.command;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
 import duke.task.Task;
-import duke.task.ToDo;
 
 import java.io.IOException;
