+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
 import duke.exception.DukeOutOfBoundsException;
 import duke.task.Task;
