+++ src/main/java/duke/exception/DukeException.java
@@ -2,6 +2,6 @@ package duke.exception;
 
 public class DukeException extends Exception {
     public DukeException(String message) {
-        super("Uh oh!!! ☹" + message);
+        super("Uh oh!!! " + message);
     }
 }
