+++ src/main/java/duke/storage/Storage.java
@@ -1,8 +1,8 @@
-package duke;
+package duke.storage;
 
+import duke.task.TaskList;
 import duke.exception.DukeException;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
