+++ src/main/java/duke/ui/Ui.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.ui;
 
+import duke.task.TaskList;
 import duke.task.Task;
 import java.util.Scanner;
